Removed all but one Serial.print() instructions, added Chip ID for debug
Serial.print() instructions removed because it affected timing, now it
is only accessible with verbose mode on, remember to use for debugging
only as it affects binding with the quad. Added Chip ID for debugging.
Added A7105_Setup() at the end of the A7105_Reset() function, so now the
4 wires SPI is still enabled after a reset and it removes some bugs
regarding the A7105-arduino communication. Many thanks to
emertonom/NicholasH1 for pointing out this.