Experimental Version!
Hello everyone, Holograph here!
Presenting this Experimental version for the LeapCopter!
- New feature : Stabilisation tuning -> While in stabilisation, you can
now tune up or down your stabilisation throttle value with the up and
down keys (subject to change), saving it each time you tweak it!
- Slightly changed the position of "hand position" text : it seemed
slightly too far to the right
Why is it only an experimental version? Well, for now there's a strange
problem that causes values higher than 5 or lower than -5 to not be
saved, as they turn to 4.99999997 (wut?), otherwise, it works just fine.