Commented out parts related to flips and rudder control.
The parts relative to flips and to rudder control with keytap have been
commented out, as we are thinking of a new way to implement flips and
because rudder control was too much of a pain, as it was relatively
random. Rudder control will soon be implemented following a less random
way.